requirements.txt
openai>=1.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
